requirement.txt
pytorch == 1.4.0
numpy == 1.19.1
scipy == 1.5.2
pandas == 1.1.3
cppimport == 20.8.4.2